pakcakges

+++ bin/install/packages/keyd.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
-echo "Installing keyd keyboard remapper..."
-
-sudo pacman -S --noconfirm keyd
-sudo rm /etc/keyd/default.conf
-sudo ln -s ~/.config/keyd/default.conf /etc/keyd/default.conf
-sudo systemctl enable --now keyd
-
+pkg="keyd"
+if omarchy-pkg-missing $pkg; then
+    echo "Installing keyd keyboard remapper..."
+    sudo pacman -S --noconfirm $pkg
+    sudo rm /etc/keyd/default.conf
+    sudo ln -s ~/.config/keyd/default.conf /etc/keyd/default.conf
+    sudo systemctl enable --now keyd
+fi
